refactor: move getLogLevel fns to levels.go
