+++ setup.go
@@ -259,40 +259,6 @@ func GetSubsystems() []string {
 	return subs
 }
 
-// GetLogLevel returns the current log level for a given subsystem as a string.
-// Passing name="*" or name="" returns the defaultLevel.
-func GetLogLevel(name string) (string, error) {
-	if name == "*" || name == "" {
-		loggerMutex.RLock()
-		defLvl := defaultLevel
-		loggerMutex.RUnlock()
-		return zapcore.Level(defLvl).String(), nil
-	}
-	if lvl, ok := levels[name]; ok {
-		return zapcore.Level(LogLevel(lvl.Level())).String(), nil
-	}
-	return "", ErrNoSuchLogger
-}
-
-// GetAllLogLevels returns a map of all current log levels for all subsystems as strings.
-// The map includes a special "*" key that represents the defaultLevel.
-func GetAllLogLevels() map[string]string {
-	result := make(map[string]string, len(levels)+1)
-
-	// Add the default level with "*" key
-	loggerMutex.RLock()
-	defLvl := defaultLevel
-	loggerMutex.RUnlock()
-	result["*"] = zapcore.Level(defLvl).String()
-
-	// Add all subsystem levels
-	for name, level := range levels {
-		result[name] = zapcore.Level(LogLevel(level.Level())).String()
-	}
-
-	return result
-}
-
 func getLogger(name string) *zap.SugaredLogger {
 	loggerMutex.Lock()
 	defer loggerMutex.Unlock()
